Move `math` module to `kernel`
It is no longer used outside of the kernel.
