+++ src/kernel/math.rs
