+++ src/kernel/mod.rs
@@ -1,5 +1,6 @@
 pub mod approximation;
 pub mod geometry;
+pub mod math;
 pub mod shapes;
 pub mod topology;
 pub mod triangulation;
